+++ package.json
@@ -99,7 +99,7 @@
 		"tempy": "^0.3.0",
 		"ts-node": "^8.5.2",
 		"tsd": "^0.11.0",
-		"typescript": "^3.7.2",
+		"typescript": "^3.7.5",
 		"xo": "^0.25.3"
 	}
 }
